requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
alembic==1.12.1
python-dotenv==1.0.0
PyJWT==2.8.0
jinja2==3.1.2
itsdangerous==2.1.2
python-multipart==0.0.6
